refactor: 뷰홀더 클래스 분리
